ml-kem: use `module_lattice::encode`
Uses the shared implementation of encoding functionality provided by the
`module-lattice` crate.

This completely replaces the previous `ml_kem::encode` module, which has
been removed, and its tests transferred to `module-lattice`.
