+++ ml-kem/src/lib.rs
@@ -57,9 +57,6 @@ mod crypto;
 /// Section 4.2.1. Conversion and Compression Algorithms, Compression and decompression
 mod compress;
 
-/// Section 4.2.1. Conversion and Compression Algorithms, Encoding and decoding
-mod encode;
-
 /// Section 5. The K-PKE Component Scheme
 mod pke;
 
@@ -78,7 +75,8 @@ pub use array;
 pub use ml_kem_512::MlKem512Params;
 pub use ml_kem_768::MlKem768Params;
 pub use ml_kem_1024::MlKem1024Params;
-pub use param::{ArraySize, ExpandedDecapsulationKey, ParameterSet};
+pub use module_lattice::encode::ArraySize;
+pub use param::{ExpandedDecapsulationKey, ParameterSet};
 pub use traits::*;
 
 use array::{
