+++ module-lattice/tests/encode.rs
@@ -1,15 +1,138 @@
 //! Tests for the `encode` module.
 
-use hybrid_array::typenum::{U1, U4, U10, U12};
-use module_lattice::algebra::{Elem, NttPolynomial, NttVector, Polynomial, Vector};
-use module_lattice::encode::{Encode, byte_decode, byte_encode};
+#![allow(clippy::integer_division_remainder_used)]
+
+use array::sizes::U3;
+use array::typenum::{Mod, Zero};
+use array::{
+    Array,
+    sizes::{U1, U2, U4, U5, U6, U8, U10, U11, U12, U256},
+};
+use getrandom::{
+    SysRng,
+    rand_core::{Rng, UnwrapErr},
+};
+use module_lattice::encode::EncodedVector;
+use module_lattice::{
+    algebra::{Elem, Field, NttPolynomial, NttVector, Polynomial, Vector},
+    encode::{ArraySize, Encode, EncodedPolynomial, EncodingSize, byte_decode, byte_encode},
+};
+use std::fmt::Debug;
+use std::ops::Rem;
 
 // Field used by ML-KEM.
 module_lattice::define_field!(KyberField, u16, u32, u64, 3329);
 
+type Int = u16;
+type DecodedValue = module_lattice::encode::DecodedValue<KyberField>;
+
+/// A helper trait to construct larger arrays by repeating smaller ones
+trait Repeat<T: Clone, D: ArraySize> {
+    fn repeat(&self) -> Array<T, D>;
+}
+
+impl<T, N, D> Repeat<T, D> for Array<T, N>
+where
+    N: ArraySize,
+    T: Clone,
+    D: ArraySize + Rem<N>,
+    Mod<D, N>: Zero,
+{
+    #[allow(clippy::integer_division_remainder_used)]
+    fn repeat(&self) -> Array<T, D> {
+        Array::from_fn(|i| self[i % N::USIZE].clone())
+    }
+}
+
 // ========================================
-// byte_encode / byte_decode round-trip tests
+// byte_encode / byte_decode tests
 // ========================================
+
+#[allow(clippy::integer_division_remainder_used)]
+fn byte_codec_test<D>(decoded: &DecodedValue, encoded: &EncodedPolynomial<D>)
+where
+    D: EncodingSize,
+{
+    // Test known answer
+    let actual_encoded = byte_encode::<KyberField, D>(decoded);
+    assert_eq!(&actual_encoded, encoded);
+
+    let actual_decoded = byte_decode::<KyberField, D>(encoded);
+    assert_eq!(&actual_decoded, decoded);
+
+    // Test random decode/encode and encode/decode round trips
+    let mut rng = UnwrapErr(SysRng);
+    let decoded = Array::<Int, U256>::from_fn(|_| (rng.next_u32() & 0xFFFF) as Int);
+    let m = match D::USIZE {
+        12 => KyberField::Q,
+        d => (1 as Int) << d,
+    };
+    let decoded = decoded.iter().map(|x| Elem::new(x % m)).collect();
+
+    let actual_encoded = byte_encode::<KyberField, D>(&decoded);
+    let actual_decoded = byte_decode::<KyberField, D>(&actual_encoded);
+    assert_eq!(actual_decoded, decoded);
+
+    let actual_reencoded = byte_encode::<KyberField, D>(&decoded);
+    assert_eq!(actual_reencoded, actual_encoded);
+}
+
+#[test]
+fn byte_codec() {
+    // The 1-bit can only represent decoded values equal to 0 or 1.
+    let decoded: DecodedValue = Array::<_, U2>([Elem::new(0), Elem::new(1)]).repeat();
+    let encoded: EncodedPolynomial<U1> = Array([0xaa; 32]);
+    byte_codec_test::<U1>(&decoded, &encoded);
+
+    // For other codec widths, we use a standard sequence
+    let decoded: DecodedValue = Array::<_, U8>([
+        Elem::new(0),
+        Elem::new(1),
+        Elem::new(2),
+        Elem::new(3),
+        Elem::new(4),
+        Elem::new(5),
+        Elem::new(6),
+        Elem::new(7),
+    ])
+    .repeat();
+
+    let encoded: EncodedPolynomial<U4> = Array::<_, U4>([0x10, 0x32, 0x54, 0x76]).repeat();
+    byte_codec_test::<U4>(&decoded, &encoded);
+
+    let encoded: EncodedPolynomial<U5> = Array::<_, U5>([0x20, 0x88, 0x41, 0x8a, 0x39]).repeat();
+    byte_codec_test::<U5>(&decoded, &encoded);
+
+    let encoded: EncodedPolynomial<U6> =
+        Array::<_, U6>([0x40, 0x20, 0x0c, 0x44, 0x61, 0x1c]).repeat();
+    byte_codec_test::<U6>(&decoded, &encoded);
+
+    let encoded: EncodedPolynomial<U10> =
+        Array::<_, U10>([0x00, 0x04, 0x20, 0xc0, 0x00, 0x04, 0x14, 0x60, 0xc0, 0x01]).repeat();
+    byte_codec_test::<U10>(&decoded, &encoded);
+
+    let encoded: EncodedPolynomial<U11> = Array::<_, U11>([
+        0x00, 0x08, 0x80, 0x00, 0x06, 0x40, 0x80, 0x02, 0x18, 0xe0, 0x00,
+    ])
+    .repeat();
+    byte_codec_test::<U11>(&decoded, &encoded);
+
+    let encoded: EncodedPolynomial<U12> = Array::<_, U12>([
+        0x00, 0x10, 0x00, 0x02, 0x30, 0x00, 0x04, 0x50, 0x00, 0x06, 0x70, 0x00,
+    ])
+    .repeat();
+    byte_codec_test::<U12>(&decoded, &encoded);
+}
+
+#[test]
+fn byte_codec_12_mod() {
+    // DecodeBytes_12 is required to reduce mod q
+    let encoded: EncodedPolynomial<U12> = Array([0xff; 384]);
+    let decoded: DecodedValue = Array([Elem::new(0xfff % KyberField::Q); 256]);
+
+    let actual_decoded = byte_decode::<KyberField, U12>(&encoded);
+    assert_eq!(actual_decoded, decoded);
+}
 
 #[test]
 fn byte_encode_decode_d1_roundtrip() {
@@ -136,9 +259,51 @@ fn polynomial_encode_decode_d12() {
 // Vector encoding tests
 // ========================================
 
+fn vector_codec_known_answer_test<D, T>(decoded: &T, encoded: &Array<u8, T::EncodedSize>)
+where
+    D: EncodingSize,
+    T: Encode<D> + PartialEq + Debug,
+{
+    let actual_encoded = decoded.encode();
+    assert_eq!(&actual_encoded, encoded);
+
+    let actual_decoded: T = Encode::decode(encoded);
+    assert_eq!(&actual_decoded, decoded);
+}
+
+#[test]
+fn vector_codec() {
+    let poly = Polynomial::new(
+        Array::<_, U8>([
+            Elem::new(0),
+            Elem::new(1),
+            Elem::new(2),
+            Elem::new(3),
+            Elem::new(4),
+            Elem::new(5),
+            Elem::new(6),
+            Elem::new(7),
+        ])
+        .repeat(),
+    );
+
+    // The required vector sizes are 2, 3, and 4.
+    let decoded: Vector<KyberField, U2> = Vector::new(Array([poly, poly]));
+    let encoded: EncodedVector<U5, U2> = Array::<_, U5>([0x20, 0x88, 0x41, 0x8a, 0x39]).repeat();
+    vector_codec_known_answer_test::<U5, Vector<KyberField, U2>>(&decoded, &encoded);
+
+    let decoded: Vector<KyberField, U3> = Vector::new(Array([poly, poly, poly]));
+    let encoded: EncodedVector<U5, U3> = Array::<_, U5>([0x20, 0x88, 0x41, 0x8a, 0x39]).repeat();
+    vector_codec_known_answer_test::<U5, Vector<KyberField, U3>>(&decoded, &encoded);
+
+    let decoded: Vector<KyberField, U4> = Vector::new(Array([poly, poly, poly, poly]));
+    let encoded: EncodedVector<U5, U4> = Array::<_, U5>([0x20, 0x88, 0x41, 0x8a, 0x39]).repeat();
+    vector_codec_known_answer_test::<U5, Vector<KyberField, U4>>(&decoded, &encoded);
+}
+
 #[test]
 fn vector_encode_decode_roundtrip() {
-    use hybrid_array::typenum::U2;
+    use array::typenum::U2;
 
     let coeffs1: [Elem<KyberField>; 256] = core::array::from_fn(|i| Elem::new((i * 3) as u16 % 16));
     let coeffs2: [Elem<KyberField>; 256] = core::array::from_fn(|i| Elem::new((i * 5) as u16 % 16));
@@ -186,7 +351,7 @@ fn ntt_polynomial_encode_decode_d12() {
 
 #[test]
 fn ntt_vector_encode_decode_roundtrip() {
-    use hybrid_array::typenum::U2;
+    use array::typenum::U2;
 
     let coeffs1: [Elem<KyberField>; 256] = core::array::from_fn(|i| Elem::new((i * 3) as u16 % 16));
     let coeffs2: [Elem<KyberField>; 256] = core::array::from_fn(|i| Elem::new((i * 5) as u16 % 16));
@@ -227,7 +392,7 @@ fn encoded_polynomial_size_d12() {
 
 #[test]
 fn encoded_vector_size() {
-    use hybrid_array::typenum::U3;
+    use array::typenum::U3;
 
     // D=4, K=3: 128 bytes per polynomial * 3 = 384 bytes
     let coeffs = [Elem::<KyberField>::new(0); 256];
